Fail build if job does not contain 4 BuildChooserSetting lines
Must be called with -Dconfig.file=../config.xml in Jenkins job

JENKINS-33695 detected that git plugin 2.4.3 no longer allowed
multiple BuildChooserSetting values in a single job configuration file.
The problem was visible when a job was loaded at Jenkins startup or
reloaded from disc.

The problem was not visible when the job configuration was saved in a
running Jenkins session without reloading the configuration from disc.