Report git commit messages from last 15 minutes
JENKINS-41906 reported that commits on another branch were causing builds
on the master branch when the web hook was invoked.  This will be used on
the master branch to confirm that if a build happens on the master branch,
it was either preceded by a git commit or is the first build of the job.