Fail if more than one git remote is defined
The bug is testing polling from the Jenkins server.  If more than
one remote is defined, there is a risk that the additional remote
has a post-receive hook script which will detect changes before
the polling detects them.  Since this bug is checking a polling
case, don't risk subverting the polling with a notifyCommit hook.