Add installation name to 'ant-latest' value
Seems to be the second time that the documentation is incorrectly
claiming that a parameter is optional when it is mandatory.

Earlier problem was with the milestone step.  Docs say that the
ordinal is optional yet if I use parentheses in the call, then the
ordinal name must appear and have a value.