Do not rely on 'git -C' when submodule not yet updated
A little too tricky for me as a reader to have the output of 'git -C'
be for the parent git repository when the submodule has not been
updated, then to be for the submodule when the submodule has been
updated. Be less tricky and make it clear which command is being
called.