Require git 1.8 or git 1.7
Problem is not visible on the newer versions of command line git that
are included with more recent operating systems.

Command line git 1.8 definitely shows the problem on CentOS 7.
Command line git 1.7 is included in the test to see if the issue is
specific to 1.8 or is connected to versions 1.8 and before.